Moved utility functions from ensemble/forest to utils
_get_n_samples() and _generate_sample_indices() previously resided in ensemble/_forest.py, leading to a circular definition
Moved them both to utils/__init__.py so they could be referenced by ensemble/_forest and tree/_class
Ensured all unit tests still pass
